+++ examples/blueprint_example/src/example_blueprint_main.py
@@ -32,12 +32,16 @@ def blueprint_input():
             output_dir_name="output",
             venv=os.environ.get(
                 "BLUEPRINT_VENV_PATH", "/home/runner/work/yotse/yotse/blueprint_venv"
-            ),
+            ),  # the environment where you installed all dependencies for your script
+            slurm_venv="~/yotse_venv",  # the environment where you installed yotse
             num_nodes=2,
             alloc_time="01:00:00",
             slurm_args=["--exclusive"],
             qcg_cfg={"log_level": "DEBUG"},
-            modules=["2021", "Python/3.9.5-GCCcore-10.3.0"],
+            modules=[
+                "2022",
+                "Python/3.10.4-GCCcore-11.3.0",
+            ],  # These are not needed if you set a slurm_venv
         ),
         parameters=[
             Parameter(